War Draft
A draft of the main War game, where the game essentially works. Decks
are made and shuffled. Cards are removed from the decks to be played,
and eventually emptied. When a player wins the cards they are added to a
queue of winnings, and that part works correctly.
The queue is still unlinked to the stack, so that needs to be linked up
for winnings to be used to continue the game.

There is also still a bug in the shuffle where a particular card will
get duplicated during the SHUFFLE (and the duplicated card is different
every time).